Refactor Dockerfile and update watchdog
The function being deployed from the store is broken.  The error is `"token" parameter not in video info for unknown reason` and the cause is well documented.  A fix has been applied and it appears as though the apk package that was in use is trailing other distribution mechanisms.

This change swaps to the version distributed by pip.  It also updates the watchdog to use the Docker image version.  Also updates the stack fileto a add the ext label.

Signed-off-by: Richard Gee <richard@technologee.co.uk>